+++ src/app/store/components/product-list-contents-display/store-product-list-contents-display.component.ts
@@ -8,7 +8,7 @@
 import { Component, EventEmitter, Inject, Input, OnDestroy, OnInit, Output } from '@angular/core';
 import { PageEvent } from '@angular/material/paginator';
 import { Observable, ReplaySubject, Subscription } from 'rxjs';
-import { map, tap } from 'rxjs/operators';
+import { finalize, map, tap } from 'rxjs/operators';
 import { API_SERVICE_INJECTION_TOKENS } from 'src/app/api/api-service-injection-tokens';
 import { ITransactionalProductListContentsDataApiService } from 'src/app/api/transactional-product-list-contents.data.api.iservice';
 import { DataPage } from 'src/models/DataPage';
@@ -24,7 +24,6 @@ import { StoreCatalogService } from '../../routes/catalog/store-catalog.service'
 export class StoreProductListContentsDisplayComponent
   implements OnInit, OnDestroy {
 
-  private loadingSubscription: Subscription;
   private pageSource = new ReplaySubject<DataPage<Product>>(1);
 
   @Input() list = new ProductList();
@@ -35,8 +34,10 @@ export class StoreProductListContentsDisplayComponent
   pageSize = 10;
   page$ = this.pageSource.asObservable();
 
+  loadingProducts = true;
   products$: Observable<Product[]>;
   totalCount$: Observable<number>;
+  loadingSubscription: Subscription;
 
   constructor(
     @Inject(API_SERVICE_INJECTION_TOKENS.dataProductLists) private productListApiService: ITransactionalProductListContentsDataApiService,
@@ -69,10 +70,12 @@ export class StoreProductListContentsDisplayComponent
   }
 
   private reloadItems() {
+    this.loadingProducts = true;
     this.loadingSubscription?.unsubscribe();
 
     this.loadingSubscription = this.productListApiService.fetchContents(this.list, this.pageIndex, this.pageSize).pipe(
-      tap(page => this.pageSource.next(page))
+      tap(page => this.pageSource.next(page)),
+      finalize(() => (this.loadingProducts = false))
     ).subscribe();
   }